+++ chift/openapi/models.py
@@ -16,6 +16,10 @@ from .openapi import (
     BackboneCommonModelsPosCommonProductCategoryItem,
     BalanceItemOut,
     BankAccountItemOut,
+    BankingAccountItem,
+    BankingCounterPartItem,
+    BankingFinancialInstitutionItem,
+    BankingTransactionItem,
     CategoryItem,
     ClientItemOut,
     ClosureItem,
@@ -208,6 +212,14 @@ class Consumer(ConsumerItem, extra="allow"):
         )  # avoid circular import
 
         return PaymentRouter(self.consumerid, self.connectionid)
+
+    @property
+    def banking(self):
+        from chift.models.consumers.banking import (
+            BankingRouter,
+        )  # avoid circular import
+
+        return BankingRouter(self.consumerid, self.connectionid)
 
     @property
     def custom(self):
@@ -566,6 +578,29 @@ class PaymentPayment(PaymentItemOut):
 
 
 class PaymentRefund(RefundItemOut):
+    pass
+
+
+# Banking
+
+
+class BankingFinancialInstitution(BankingFinancialInstitutionItem):
+    pass
+
+
+class BankingAccount(BankingAccountItem):
+    pass
+
+
+class BankingTransaction(BankingTransactionItem):
+    pass
+
+
+class BankingCounterpart(BankingCounterPartItem):
+    pass
+
+
+class BankingAttachment(AttachmentItemOut):
     pass